Reproduce race condition better in test case
Previously, the test case would succeed even when a command queue was
being used. Since that leads to bugs on a real CCU (where the time for
sending a command is not negligible), this commit introduces a similar
communication delay to the test.

At the same time, this allows us to reduce the number of iterations that
this test needs to run in order to be reasonably sure that a race
condition is absent.